requirements-slack.txt
python-slackclient
